src/backend/content-management-service/requirements.txt
flask
flask-pymongo
python-dotenv
Flask-Cors
flask_marshmallow